+++ packages/flutter_sangfor/lib/flutter_sangfor.dart
@@ -20,6 +20,7 @@ export 'src/fd_packet_device.dart';
 export 'src/utun_device.dart';
 export 'src/android_vpn.dart';
 export 'src/ios_vpn.dart';
+export 'src/ohos_vpn.dart';
 
 /// Controls an aTrust-compatible VPN session.
 class FlutterSangfor {